mnist_tensorflow_dnn/requirement.txt
tensoflow 1.4.0Package          Version
---------------- -------
absl-py          0.8.0  
astor            0.8.0  
backcall         0.1.0  
decorator        4.4.0  
gast             0.2.2  
grpcio           1.23.0 
ipython          7.7.0  
ipython-genutils 0.2.0  
jedi             0.15.1 
Markdown         3.1.1  
numpy            1.14.5 
parso            0.5.1  
pexpect          4.7.0  
pickleshare      0.7.5  
pip              19.2.3 
pkg-resources    0.0.0  
prompt-toolkit   2.0.9  
protobuf         3.9.1  
ptyprocess       0.6.0  
Pygments         2.4.2  
setuptools       39.1.0 
six              1.12.0 
tensorboard      1.10.0 
tensorflow-gpu   1.10.0 
termcolor        1.1.0  
traitlets        4.3.2  
wcwidth          0.1.7  
Werkzeug         0.15.5 
wheel            0.33.6 
